Make the split out plugins build everywhere.
This seems to break pytest_* tests, and so may need to be reverted. This works
as expected with the C++ tests as far as I can tell. Initially it did not work
as expected due to the setting of OUTPUT_DIRECTORY. This is set globally, and
should not be set for individual targets.


git-svn-id: https://openbabel.svn.sourceforge.net/svnroot/openbabel/openbabel/trunk@4182 71e0b675-c112-0410-beda-96664073b5c2