Added environment setting commands to pybel build and install commands.
This fix passes CFLAGS set in CMake to setup.py. setup.py is also passed "-I${EIGEN2_INCLUDE_DIR}", which will otherwise not compile if the Eigen2 headers are not placed in a default location.

This fix uses the export environment setting command, which will not work if the shell does not support this ( sh or bash do, but not csh). A better fix is probably in order



git-svn-id: https://openbabel.svn.sourceforge.net/svnroot/openbabel/openbabel/trunk@4350 71e0b675-c112-0410-beda-96664073b5c2